Update README.md
This program is a simple example of optimization through order of processing conditional logic. Using the task of determining if a number is prime, I will express the meaning of optimization through order of processing.

You may get right to reading the example code extracted from the program if you like.